Be consistent with class name 'Ruleset'
Originally, we agreed that 'RuleSet' is two words, so camelcase it. But then
when I went ahead and made all the tests, I used a lowercase s, and Tom agrees
with the lower case s this time. (Is he just agreeing with whatever I do?) So
change all the uses of capitalized 'RuleSet' to the singularly capitalized
'Ruleset'